+++ apps/insights/src/get-host.ts
