+++ apps/insights/src/server/pyth.ts
@@ -1,10 +1,10 @@
 import { parse } from "superjson";
 import { z } from "zod";
 
+import { DEFAULT_CACHE_TTL } from "../cache";
 import { VERCEL_REQUEST_HEADERS } from "../config/server";
+import { getHost } from "../get-host";
 import { Cluster, ClusterToName, priceFeedsSchema } from "../services/pyth";
-import { DEFAULT_CACHE_TTL } from "../utils/cache";
-import { getHost } from "../utils/get-host";
 
 export async function getPublishersForFeedRequest(
   cluster: Cluster,
